requirements.txt
aniso8601==9.0.1
async-timeout==4.0.2
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.1
loguru==0.6.0
MarkupSafe==2.1.2
peewee==3.16.0
PyMySQL==1.0.2
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.1
six==1.16.0
Werkzeug==2.2.3
win32-setctime==1.1.0
